Added two things that could be used to compare with multiple draw call performance without having to implement it.
The first is drawing (direct eval) each patch individually instead of in just one draw call. To simulate having to switch descriptor sets for the next patch it just switches to the same each time. The other thing is adding dummy data to the control point buffer to see if the size of the shader storage buffers has any cost for rendering.